update presenter y view
