+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".moredetails.fulllogic.view.OtherInfoViewActivity"
+        <activity android:name=".moredetails.mvp.presenter.OtherInfoViewActivity"
            android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
