--- app/src/main/kotlin/ayds/lisboa/songinfo/moredetails/fulllogic/view/OtherInfoEvents.kt
@@ -1,5 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.fulllogic.view
-
-sealed class OtherInfoEvent {
-    object OpenInfoUrl: OtherInfoEvent()
-}
