+++ app/src/main/kotlin/ayds/lisboa/songinfo/moredetails/mvp/data/external/LastFmToArtistInfoResolver.kt
@@ -1,9 +1,9 @@
-package ayds.lisboa.songinfo.moredetails.fulllogic.data.external
+package ayds.lisboa.songinfo.moredetails.mvp.data.external
 
-import ayds.lisboa.songinfo.moredetails.fulllogic.domain.ArtistInfo
+import ayds.lisboa.songinfo.moredetails.mvp.domain.ArtistInfo
 import com.google.gson.Gson
 import com.google.gson.JsonObject
-import ayds.lisboa.songinfo.moredetails.fulllogic.domain.ArtistInfo.LastFmArtistInfo
+import ayds.lisboa.songinfo.moredetails.mvp.domain.ArtistInfo.LastFmArtistInfo
 
 private const val JSON_ARTIST = "artist"
 private const val JSON_BIO = "bio"
